[reformat][antom-sdk-automation] automated change

+++ com/alipay/api/response/billing/AlipayMeterUploadEventResponse.go
@@ -8,6 +8,6 @@ import (
 type AlipayMeterUploadEventResponse struct {
 	response.AlipayResponse
 	Result     *model.Result  `json:"result,omitempty"`
-	RetryAfter int32          `json:"retryAfter,omitempty"`
+	RetryAfter int64          `json:"retryAfter,omitempty"`
 	Errors     []*model.Error `json:"errors,omitempty"`
 }
